Split ReceiverController: extract settings commands
Fourth in the receiver-split arc (#809 cleanup loop, #815
peer-link sessions, #821 identity commands). Extracts the
receiver-side settings WS commands plus the shared
read-modify-write helpers into a new sibling module
controllers/remote_build/settings_receiver.py:

- get_settings - returns the wire view of receiver-side
  settings (master enabled toggle + cleanup TTL + peers).
- set_settings - persists the master enabled toggle +
  optional cleanup_ttl_seconds with strict-bool validation
  and live-rebind of the peer-link listener.
- to_view - shared projection helper used by both commands
  and by the post-mutation response on peer-CRUD WS commands.
- modify_settings - read-modify-write under the metadata
  lock, returns the wire view.
- current_settings_view - convenience wrapper that loads +
  projects (used by approve_peer / remove_peer / set_settings).

The two @api_command-decorated WS methods stay as bound
delegates on the controller (WS dispatch needs them on the
class). The three private helpers (_to_view, _modify_settings,
_current_settings_view) also stay as bound delegates because
the still-on-controller peer-CRUD methods (approve_peer,
remove_peer) and pair-flow methods (record_pair_request,
_lookup_peer_response) call them.

Module-named settings_receiver to disambiguate from the
existing settings_commands.py which holds the offloader-side
settings WS commands.

Imports moved out of receiver.py: collections.abc.Callable,
remote_build_settings_transaction, MAX_CLEANUP_TTL_SECONDS,
MIN_CLEANUP_TTL_SECONDS.

Zero test changes - the @api_command surface and bound
delegates resolve unchanged.

Receiver: 789 -> 713 lines (-76). Combined with #809 +
#815 + #821: 1083 -> 713 (-370, -34%). All 3417 tests pass.
