+++ esphome_device_builder/controllers/remote_build/receiver.py
@@ -28,8 +28,6 @@ from ...helpers.api import CommandError, api_command
 from ...helpers.event_bus import Event
 from ...helpers.storage import Store
 from ...models import (
-    MAX_CLEANUP_TTL_SECONDS,
-    MIN_CLEANUP_TTL_SECONDS,
     TERMINAL_JOB_EVENTS,
     ErrorCode,
     EventType,
@@ -48,9 +46,8 @@ from ...models import (
 )
 from ..config import (
     load_remote_build_settings,
-    remote_build_settings_transaction,
 )
-from . import cleanup_loop, identity_commands, peer_link_sessions
+from . import cleanup_loop, identity_commands, peer_link_sessions, settings_receiver
 from ._receiver_state import ReceiverState
 from ._shared import _RemoteBuildBase, drain_tasks
 from ._storage_codecs import (
@@ -215,23 +212,11 @@ class ReceiverController(_RemoteBuildBase):  # noqa: PLR0904
     @api_command("remote_build/get_settings")
     async def get_settings(self, **kwargs: Any) -> RemoteBuildSettingsView:
         """Return the receiver-side remote-build settings (wire view)."""
-        return self._to_view(await self._load_settings_async())
+        return await settings_receiver.get_settings(self)
 
     def _to_view(self, settings: RemoteBuildSettings) -> RemoteBuildSettingsView:
-        """Project receiver settings to wire view, merging in-memory peers.
-
-        The peer list is RAM-canonical: PENDING entries live in
-        ``self.state.pending_peers`` for the active pairing window's
-        lifetime (never hit disk) and APPROVED entries live in
-        ``self.state.approved_peers`` / its per-file ``Store``.
-        ``settings`` is consulted for the master ``enabled``
-        toggle.
-        """
-        return RemoteBuildSettingsView(
-            enabled=settings.enabled,
-            cleanup_ttl_seconds=settings.cleanup_ttl_seconds,
-            peers=self._peer_summaries(),
-        )
+        """Project receiver settings to wire view, merging in-memory peers."""
+        return settings_receiver.to_view(self, settings)
 
     def _peer_summaries(self) -> list[PeerSummary]:
         """Merge PENDING + APPROVED into a single ``PeerSummary`` list.
@@ -262,33 +247,8 @@ class ReceiverController(_RemoteBuildBase):  # noqa: PLR0904
     async def _modify_settings(
         self, mutator: Callable[[RemoteBuildSettings], None]
     ) -> RemoteBuildSettingsView:
-        """
-        Run ``mutator`` against the current settings and persist the result.
-
-        Wraps :func:`remote_build_settings_transaction` so the
-        whole read-modify-write happens under the metadata lock,
-        so two concurrent callers can't both read the same starting
-        value and have the second save wipe the first's change.
-        Runs in the default executor since the transaction does
-        blocking JSON I/O. Returns the wire view so the response
-        leaving this method can never carry ``secret_sha256``.
-
-        ``mutator`` is invoked with the freshly-loaded settings
-        and is expected to mutate it in place. A
-        :class:`CommandError` raised inside the mutator (e.g.
-        duplicate-detection on add) propagates out and discards
-        the pending write; same exception-on-discard contract as
-        :func:`metadata_transaction`.
-        """
-
-        def _txn() -> RemoteBuildSettings:
-            with remote_build_settings_transaction(self._db.settings.config_dir) as settings:
-                mutator(settings)
-                return settings
-
-        loop = asyncio.get_running_loop()
-        settings = await loop.run_in_executor(None, _txn)
-        return self._to_view(settings)
+        """Run *mutator* against the current settings and persist the result."""
+        return await settings_receiver.modify_settings(self, mutator)
 
     @api_command("remote_build/set_settings")
     async def set_settings(
@@ -298,46 +258,10 @@ class ReceiverController(_RemoteBuildBase):  # noqa: PLR0904
         cleanup_ttl_seconds: int | None = None,
         **kwargs: Any,
     ) -> RemoteBuildSettingsView:
-        """
-        Persist the receiver-side ``enabled`` master switch.
-
-        Read-modify-write so peers / other fields stay intact.
-        Strict-bool validation defeats truthiness coercion on
-        this security-sensitive toggle.
-
-        Optional ``cleanup_ttl_seconds`` updates the cleanup
-        sweep threshold, range-checked against
-        :data:`MIN_CLEANUP_TTL_SECONDS` /
-        :data:`MAX_CLEANUP_TTL_SECONDS`. Omit to keep current.
-
-        Live-rebinds the peer-link listener: True runs the
-        startup bind path, False tears down + clears the mDNS
-        pin/port advertise. Fail-soft on bind error.
-        """
-        if not isinstance(enabled, bool):
-            msg = "remote_build/set_settings: 'enabled' must be a boolean"
-            raise CommandError(ErrorCode.INVALID_ARGS, msg)
-        if cleanup_ttl_seconds is not None:
-            # bool subclasses int, so reject ``True`` first to
-            # avoid a misleading OUT_OF_RANGE on a type error.
-            if isinstance(cleanup_ttl_seconds, bool) or not isinstance(cleanup_ttl_seconds, int):
-                msg = "remote_build/set_settings: 'cleanup_ttl_seconds' must be an integer"
-                raise CommandError(ErrorCode.INVALID_ARGS, msg)
-            if not MIN_CLEANUP_TTL_SECONDS <= cleanup_ttl_seconds <= MAX_CLEANUP_TTL_SECONDS:
-                msg = (
-                    f"remote_build/set_settings: 'cleanup_ttl_seconds' must be between "
-                    f"{MIN_CLEANUP_TTL_SECONDS} and {MAX_CLEANUP_TTL_SECONDS}"
-                )
-                raise CommandError(ErrorCode.INVALID_ARGS, msg)
-
-        def _set(settings: RemoteBuildSettings) -> None:
-            settings.enabled = enabled
-            if cleanup_ttl_seconds is not None:
-                settings.cleanup_ttl_seconds = cleanup_ttl_seconds
-
-        view = await self._modify_settings(_set)
-        await self._db.apply_remote_build_enabled()
-        return view
+        """Persist the receiver-side ``enabled`` master switch."""
+        return await settings_receiver.set_settings(
+            self, enabled=enabled, cleanup_ttl_seconds=cleanup_ttl_seconds
+        )
 
     # ------------------------------------------------------------------
     # Offloader-side settings: master toggle + per-pairing enable.
@@ -439,7 +363,7 @@ class ReceiverController(_RemoteBuildBase):  # noqa: PLR0904
 
     async def _current_settings_view(self) -> RemoteBuildSettingsView:
         """Load settings from disk and project to the wire view (post-mutation response)."""
-        return self._to_view(await self._load_settings_async())
+        return await settings_receiver.current_settings_view(self)
 
     # ------------------------------------------------------------------
     # Peer-link Noise WS dispatch helpers — called by the post-handshake
